+++ segmentation/include/pcl/segmentation/impl/organized_connected_component_segmentation.hpp
@@ -125,7 +125,7 @@ pcl::OrganizedConnectedComponentSegmentation<PointT, PointLT>::segment (pcl::Poi
   labels.points.resize (input_->points.size (), invalid_pt);
   labels.width = input_->width;
   labels.height = input_->height;
-  unsigned int clust_id = 0;
+  size_t clust_id = 0;
   
   //First pixel
   if (std::isfinite (input_->points[0].x))
@@ -205,8 +205,8 @@ pcl::OrganizedConnectedComponentSegmentation<PointT, PointLT>::segment (pcl::Poi
   }
   
   std::vector<unsigned> map (clust_id);
-  unsigned max_id = 0;
-  for (unsigned runIdx = 0; runIdx < run_ids.size (); ++runIdx)
+  size_t max_id = 0;
+  for (size_t runIdx = 0; runIdx < run_ids.size (); ++runIdx)
   {
     // if it is its own root -> new region
     if (run_ids[runIdx] == runIdx)
@@ -216,7 +216,7 @@ pcl::OrganizedConnectedComponentSegmentation<PointT, PointLT>::segment (pcl::Poi
   }
 
   label_indices.resize (max_id + 1);
-  for (unsigned idx = 0; idx < input_->points.size (); idx++)
+  for (size_t idx = 0; idx < input_->points.size (); idx++)
   {
     if (labels[idx].label != invalid_label)
     {
